Restructured import code to make it more modular
