+++ lib/import/vocab_term_importer.rb
@@ -1,0 +1,13 @@
+# Imports a spreadsheet containing vocab terms
+module Import
+  module VocabTermImporter
+
+    include PublishableImporter
+    include RowParser
+
+    def import_row(row, row_number)
+
+    end
+
+  end
+end
