+++ lib/tasks/exercises/import.rake
@@ -7,10 +7,11 @@ namespace :exercises do
     #          will import exercises from exercises.xlsx and
     #          assign the user with ID 1 as the author and
     #          solution author, and the user with ID 2 as the CR holder
-    desc "import an xlsx file"
+    desc "import exercises from an xlsx file"
     task :xlsx, [:filename, :author_id, :ch_id, :skip_first_row] => :environment do |t, args|
       # Output import logging info to the console (except in the test environment)
       original_logger = Rails.logger
+
       begin
         Rails.logger = ActiveSupport::Logger.new(STDOUT) unless Rails.env.test?
 
@@ -28,10 +29,11 @@ namespace :exercises do
     #          will import exercises from the first xlsx file found within exercises.zip and
     #          assign the user with ID 1 as the author and
     #          solution author, and the user with ID 2 as the CR holder
-    desc "import a zip file"
+    desc "import exercises from a zip file"
     task :zip, [:filename, :author_id, :ch_id, :skip_first_row] => :environment do |t, args|
       # Output import logging info to the console (except in the test environment)
       original_logger = Rails.logger
+
       begin
         Rails.logger = ActiveSupport::Logger.new(STDOUT) unless Rails.env.test?
 
@@ -47,10 +49,11 @@ namespace :exercises do
     # filename, book_name, book_archive_url
     # Example: rake exercises:import:qb[qb.json,physics,https://archive.cnx.org/contents/031da8d3-b525-429c-80cf-6c8ed997733a]
     #          will import exercises from qb.json for the college physics book
-    desc "import a quadbase json file"
+    desc "import exercises from a quadbase json file"
     task :qb, [:filename, :book_name, :book_archive_url] => :environment do |t, args|
       # Output import logging info to the console (except in the test environment)
       original_logger = Rails.logger
+
       begin
         Rails.logger = ActiveSupport::Logger.new(STDOUT) unless Rails.env.test?
