+++ lib/tasks/exercises/import_old.rake
@@ -8,10 +8,11 @@ namespace :exercises do
       #          will import exercises from exercises.xlsx and
       #          assign the user with ID 1 as the author and
       #          solution author, and the user with ID 2 as the CR holder
-      desc "import an xlsx file"
+      desc "import exercises from an xlsx file (old format)"
       task :xlsx, [:filename, :author_id, :ch_id, :skip_first_row] => :environment do |t, args|
         # Output import logging info to the console (except in the test environment)
         original_logger = Rails.logger
+
         begin
           Rails.logger = ActiveSupport::Logger.new(STDOUT) unless Rails.env.test?
 
@@ -29,10 +30,11 @@ namespace :exercises do
       #          will import exercises from the first xlsx file found within exercises.zip and
       #          assign the user with ID 1 as the author and
       #          solution author, and the user with ID 2 as the CR holder
-      desc "import a zip file"
+      desc "import exercises from a zip file (old format)"
       task :zip, [:filename, :author_id, :ch_id, :skip_first_row] => :environment do |t, args|
         # Output import logging info to the console (except in the test environment)
         original_logger = Rails.logger
+
         begin
           Rails.logger = ActiveSupport::Logger.new(STDOUT) unless Rails.env.test?
