+++ lib/tasks/exercises/tag.rake
@@ -9,6 +9,7 @@ namespace :exercises do
     task :xlsx, [:filename, :skip_first_row] => :environment do |t, args|
       # Output import logging info to the console (except in the test environment)
       original_logger = Rails.logger
+
       begin
         Rails.logger = ActiveSupport::Logger.new(STDOUT) unless Rails.env.test?
